Updates to GUI
Commented out one of the extra methods for CNN

Also added an example image loader for the new images tab. Seizure warning though, loads random images really fast. Will tone this down in future.

Next priority is getting a moving ROI bound to the hand that is compat with CNN and also removing the arm as it shifts the centroid.