profile settings expanded,
reworked self.accelregions to be a part of the user profile dictionary.

Added accel_speeds to config and dictionary to control the acceleration of mouse within the accel_regions.

Added comments for imports of unfamiliar module "ast", used for taking the literal list out of config.

Partially implemented accel_regions in modify settings, however, setting is not currently added. because using mock profiles to show their effect.

Bug fix: When changing the pointer to index, didn't consider its impact on the roi for classification. Amended this issue with.